.claude/skills/fastapi-builder/assets/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.16
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
alembic>=1.13.1
pytest>=7.4.0
httpx>=0.25.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9  # For PostgreSQL
asyncpg>=0.29.0          # For async PostgreSQL
pymysql>=1.1.1           # For MySQL
cryptography>=41.0.8     # For security features
slowapi>=0.1.9          # For rate limiting (optional)
redis>=5.0.1            # For caching (optional)
celery>=5.3.0           # For background tasks (optional)